#91: implemented application context containing the users identity
