+++ src/webui/App.tsx
@@ -1,0 +1,12 @@
+
+import React from 'react';
+
+import AppContext from './AppContext';
+import Routes from './Routes';
+
+export default function App()
+{
+    return <AppContext>
+        <Routes />
+    </AppContext>;
+}
