+++ src/webui/pages/templates/Application.tsx
@@ -1,23 +1,22 @@
 
 import React from 'react';
 
-import ImageView from '../../../domain/image/view/ImageView';
-import CreatorView from '../../../domain/creator/view/CreatorView';
+import type CreatorView from '../../../domain/creator/view/CreatorView';
 
 import { ApplicationSidebar } from '../../components/module';
 import { ApplicationLayout } from '../../layouts/module';
+
+import { useAppContext } from '../../AppContext';
 
 export type Props = {
     children: React.ReactNode;
 };
 
-const IMAGE_URL = 'https://masking.tech/images/peter.jpg';
-
-const portrait = new ImageView(IMAGE_URL);
-const identity = new CreatorView('0', 'Peter van Vliet', 'petermasking', portrait, new Date(), 0, 0, 0);
-
 export default function Template({ children }: Props)
 {
+    const context = useAppContext();
+    const identity = context.identity as CreatorView;
+
     const sidebar = <ApplicationSidebar identity={identity} />;
 
     return <ApplicationLayout main={children} aside={sidebar} />;
